requirements.lock.txt
aioice==0.9.0
aiortc==1.10.1
alembic==1.16.4
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
av==13.1.0
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.40.16
botocore==1.40.16
cachetools==6.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.0.0
contourpy==1.3.3
cryptography==44.0.2
cycler==0.12.1
distlib==0.3.9
Django==5.1.4
dnspython==2.7.0
entrypoints==0.4
et_xmlfile==2.0.0
Faker==37.5.3
fastapi==0.116.1
favicon==0.7.0
filelock==3.17.0
fonttools==4.59.1
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.6.0
h11==0.16.0
holamundoplayermsantanatest==0.0.1
htbuilder==0.9.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lxml==6.0.1
Mako==1.3.10
Markdown==3.8.2
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.5
narwhals==2.2.0
numpy==2.3.0
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
pdf2image==1.17.0
pillow==11.2.1
pipenv==2024.4.1
platformdirs==4.3.6
plotly==6.3.0
prometheus_client==0.22.1
protobuf==6.31.1
psycopg==3.2.9
psycopg-binary==3.2.9
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyee==12.1.1
PyJWT==2.10.1
pylibsrtp==0.11.0
pymdown-extensions==10.16.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.0
s3transfer==0.13.1
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.17.2
snowflake-snowpark-python==1.37.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.3
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.47.3
streamlit==1.48.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.6
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.35.0
validators==0.35.0
virtualenv==20.29.2
watchdog==6.0.0
wheel==0.45.1
xlrd==2.0.2
